src/app/requirements.txt
Flask==2.0.3
Flask-Classful==0.14.2
Flask-PyMongo==2.3.0
mysql-connector-python==8.0.29
fpdf2==2.5.5
pandas==1.1.5
openpyxl
datetime